Rename `api.notifications` to `api.emergencyNotifications`
